backend/requirements.txt
# Core Django dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
Pillow==10.1.0
django-filter==23.3
markdown==3.5.1
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0

# Authentication and permissions
django-allauth==0.58.2
djoser==2.2.0
django-rest-auth==0.9.5
simple-jwt==0.1.5
django-oauth-toolkit==2.3.0

# Forms and UI
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-ckeditor==6.7.0
django-colorfield==0.9.0

# Content management
django-taggit==4.0.0
django-mptt==0.14.0
django-cleanup==8.0.0
django-storages==1.14.2

# Scheduling and background tasks
django-scheduler==0.10.1
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
flower==2.0.1

# Payment processing
paypalrestsdk==1.13.1
stripe==7.3.0

# API documentation
drf-yasg==1.21.7
coreapi==2.3.3

# Performance and caching
django-redis==5.4.0
django-cacheops==7.0.2

# Monitoring and logging
sentry-sdk==1.32.0
django-debug-toolbar==4.2.0

# Security
django-axes==6.1.1
django-honeypot==1.0.1

# Utilities
django-extensions==3.2.3
django-import-export==3.3.1
django-model-utils==4.3.1

# Testing
pytest==7.4.3
pytest-django==4.5.2
factory-boy==3.3.0
coverage==7.3.2
